Fix integer type for cycle values in trace and score board painters

+++ src/qt/ball_trace_painter.cpp
@@ -183,7 +183,7 @@ BallTracePainter::drawTrace( QPainter & painter )
         {
             return false;
         }
-        start_cycle = std::max( 1L, view->time().cycle() + opt.autoTraceStart() );
+        start_cycle = std::max( 1, static_cast< int >( view->time().cycle() + opt.autoTraceStart() ) );
     }
     else if ( start_cycle >= end_cycle )
     {
