+++ src/qt/player_trace_painter.cpp
@@ -87,7 +87,7 @@ PlayerTracePainter::draw( QPainter & painter )
         {
             return;
         }
-        start_cycle = std::max( 1L, view->time().cycle() + opt.autoTraceStart() );
+        start_cycle = std::max( 1, static_cast< int >( view->time().cycle() + opt.autoTraceStart() ) );
     }
     else if ( start_cycle >= end_cycle )
     {
